docs: use information from the previous release to create pages (#4457)
