+++ Makefile
@@ -88,9 +88,13 @@ go.mod: FORCE
 	go mod verify
 go.sum: go.mod
 
-expand_website_templates:
-	go run ./scripts/expand_website_templates/main.go
-.PHONY: expand_website_templates
+website_expand_templates:
+	go run ./scripts/website/expand_templates/
+.PHONY: website_expand_templates
+
+website_dump_info:
+	go run ./scripts/website/dump_info/
+.PHONY: dump_info
 
 update_contributors_list:
 	cd .github/contributors && npm run all
